Add refactoring from first add_items branch

+++ Gasoline/Refuel.swift
@@ -28,14 +28,14 @@ public struct Refuel: Codable {
     public let createdAt: Date
 
     /// The total costs in the major curency value (Euro, Dollar, ...)
-    public var totalPrice: Double { return literPrice * fuelAmount.value }
+    public var totalPrice: Decimal { return literPrice * Decimal(fuelAmount.value) }
     /// The date the refuel action has been done
     public let date: Date
     /// The currency that applies to all price values of this class.
     public let currencyCode: String
     /// The price per liter in the major curency value (Euro, Dollar, ...)
-    public let literPrice: Double
-    /// The taken fuel amount im milli liters
+    public let literPrice: Decimal
+    /// The fuel amount in liters
     public let fuelAmount: Measurement<UnitVolume>
     /// The current mileage right at the gas station
     public let mileage: Measurement<UnitLength>
@@ -47,7 +47,7 @@ extension Refuel {
 
     init(date: Date,
          currencyCode: String = "EUR",
-         literPrice: Double,
+         literPrice: Decimal,
          fuelAmount: Measurement<UnitVolume>,
          mileage: Measurement<UnitLength>,
          note: String? = nil) {
